refactor : main 패키지를 search 패키지와 host 패키지로 리펙토링
