+++ src/main/java/com/waitit/capstone/domain/search/SearchController.java
@@ -1,4 +1,4 @@
-package com.waitit.capstone.domain.main;
+package com.waitit.capstone.domain.search;
 
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RestController;
 @AllArgsConstructor
 @RestController
 @RequestMapping("/main")
-public class MainController {
+public class SearchController {
 
-    private final MainService mainService;
+    private final SearchService searchService;
 
 }
